@@ -1,133 +1,142 @@
-import React, { Component } from 'react';
+import React, { useEffect, useState, useRef } from 'react';
 import axios from 'axios';
 import Photo from '../../components/Photo/Photo';
-import $ from 'jquery';
 import Masonry from 'masonry-layout/dist/masonry.pkgd.min.js';
 import imagesLoaded from 'imagesloaded';
 import API_KEY from '../../key';
+import { TimelineLite, CSSPlugin, gsap } from 'gsap';
 import './App.sass';
 
-class App extends Component {
-  componentWillMount() {
+const App = () => {
+  const [photos, setPhotos] = useState([]);
+  const [disabled, setDisabled] = useState(true);
+  const [text, setText] = useState("");
+
+  // browser issues and warning removed
+  gsap.registerPlugin(CSSPlugin);
+  gsap.config({
+    nullTargetWarn: false
+  });
+
+  // refs for elements to animate
+  const gridParent = useRef(null);
+  const form = useRef(null);
+  const subHeader = useRef(null);
+
+  useEffect(() => {
+    // after scrolling add a box shadow on element
+    window.onscroll = () => {
+      let findElem = document.getElementById("navSearch");
+      if (window.scrollY > 1) {
+        findElem.setAttribute("style", "box-shadow: 0 4px 5px 0 rgba(0,0,0,0.10), 0 1px 10px 0 rgba(0,0,0,0.08), 0 2px 4px -1px rgba(0,0,0,0.03);transition: all .15s linear;");
+      } else {
+        findElem.removeAttribute("style");
+      }
+    }
+
+    // animate the nav items via gsap
+    document.querySelector("#navSearch").style.visibility = "visible";
+    let tl = new TimelineLite();
+    tl.from(form.current, 1, {y: "-30px", opacity: 0, delay: 1, ease: "power3"}, 0.1)
+      .from(subHeader.current, 1, {y: "30px", opacity: 0, delay: 1, ease: "power3"}, 0.2);
+
     // fetch images from pexels via axios
     axios({
-      method: 'GET',
-      url: `https://api.pexels.com/v1/search?query=nature&per_page=50&page=1`,
+      method: "GET",
+      url: "https://api.pexels.com/v1/search?query=nature&per_page=50&page=1",
       headers: {
         Authorization: API_KEY
       }
     })
-    .then(res => {
-      this.setState({photos: res.data.photos})
-    })
-    .catch(error => {
-      console.log(error);
-    });
-  }
+      .then(res => {
+        const photos = res.data.photos;
+        setPhotos(photos);
+      })
+      .catch(error => {
+        console.log(error);
+      });
+  }, []);
 
-  componentDidMount() {
-    $(document).ready(() => {
-      // wait for all images to load then initialize the masonry layout
-      const gridContainer = document.querySelector('.grid');
-      setTimeout(() => {
-        let masonryInit;
-        imagesLoaded(gridContainer, () => {
-          masonryInit = new Masonry(gridContainer, {
-            itemSelector: '.grid-item'
-          });
-          // masonryInit.on('layoutComplete', function() {
-          //   // console.log(items.length);
-          //   masonryInit.layout();
-          // });
-          masonryInit.layout();
-        });
-      }, 600);
-      window.onscroll = () => {
-        let findElem = document.getElementById('navSearch');
-        if (window.scrollY > 1) {
-          findElem.setAttribute('style', 'box-shadow: 0 4px 5px 0 rgba(0,0,0,0.10), 0 1px 10px 0 rgba(0,0,0,0.08), 0 2px 4px -1px rgba(0,0,0,0.03);transition: all .15s linear;');
-        } else {
-          findElem.removeAttribute('style');
-        }
-      }
+  useEffect(() => {
+    // wait for all images to load then initialize the masonry layout
+    const gridContainer = document.querySelector(".grid");
+    gridContainer.style.visibility = "visible";
+    let masonryInit;
+    imagesLoaded(gridContainer, () => {
+      masonryInit = new Masonry(gridContainer, {
+        itemSelector: ".grid-item"
+      });
+      masonryInit.layout();
     });
-  }
 
-  state = {
-    photos: [],
-    disabled: true,
-    text: ''
-  }
+    // animate each image via gsap
+    let tl = new TimelineLite({onComplete: removeTransform});
+    tl.from([gridParent.current.childNodes], 1, {y: "100%", opacity: 0, delay: 1, ease: "power3", stagger: {amount: 2.5}}, 0.2);
+  }, [photos]);
 
   // check if input is empty
-  checkDisabled = (text) => {
+  const checkDisabled = text => {
     if (text.length === 0) return true;
     return false;
-  }
+  };
+
+  // remove the transform property from the style to allow gallery to work
+  const removeTransform = () => {
+    gridParent.current.childNodes.forEach(node => {
+      node.style.transform = "none";
+    });
+  };
 
   // form submit fetch new images
-  searchImage = (event) => {
+  const searchImage = event => {
     event.preventDefault();
     axios({
-      method: 'GET',
-      url: `https://api.pexels.com/v1/search?query=${this.state.text}&per_page=50&page=1`,
+      method: "GET",
+      url: `https://api.pexels.com/v1/search?query=${text}&per_page=50&page=1`,
       headers: {
         Authorization: API_KEY
       }
     })
-    .then(res => {
-      this.setState({photos: res.data.photos})
-    })
-    .catch(error => {
-      console.log(error);
-    });
-    $(document).ready(() => {
-      // wait for all images to load then initialize the masonry layout
-      const gridContainer = document.querySelector('.grid');
-      setTimeout(() => {
-        let masonryInit;
-        imagesLoaded(gridContainer, () => {
-          masonryInit = new Masonry(gridContainer, {
-            itemSelector: '.grid-item'
-          });
-          // masonryInit.on('layoutComplete', function(items) {
-          //   // console.log(items.length);
-          //   masonryInit.layout();
-          // });
-          masonryInit.layout();
-        });
-      }, 600);
-    });
+      .then(res => {
+        const photos = res.data.photos;
+        setPhotos(photos);
+      })
+      .catch(error => {
+        console.log(error);
+      });
   }
 
-  render() {
-    return (
-      <div className="container-fluid px-0">
-        <div id="navSearch" className="row m-0 fixed-top bg-white">
-          <form onSubmit={this.searchImage} className="col-12 px-2 py-3 text-center">
-            <input onChange={event => this.setState({text: event.target.value, disabled: this.checkDisabled(event.target.value)})} className="form-control search-input righteous" type="search" placeholder="Search Images" />
-            <button className="btn btn-dark search-btn righteous" disabled={this.state.disabled} type="submit">Search</button>
-          </form>
-          <div className="col-12 px-2 text-center">
-            <p style={{fontSize: '.75rem'}} className="text-muted">Powered by <a href="https://www.pexels.com/">Pexels</a></p>
-          </div>
-        </div>
-        <div style={{paddingTop: '95px'}} className="row m-0">
-          <div className="col-12 p-0 grid">
-            {this.state.photos.map(photo => (
-              <Photo
-                key={photo.id}
-                mediumImage={photo.src.medium}
-                largeImage={photo.src.large}
-                url={photo.url}
-                photographer={photo.photographer}
-              />
-            ))}
-          </div>
+  return (
+    <div className="container-fluid px-0">
+      <div id="navSearch" className="row m-0 fixed-top bg-white">
+        <form ref={form} onSubmit={searchImage} className="col-12 px-2 py-3 text-center">
+          <input
+            onChange={e => {
+              setText(e.target.value);
+              setDisabled(checkDisabled(e.target.value));
+            }}
+            className="form-control search-input righteous" type="search" placeholder="Search Images" />
+          <button className="btn btn-dark search-btn righteous" disabled={disabled} type="submit">Search</button>
+        </form>
+        <div ref={subHeader} className="col-12 px-2 text-center">
+          <p style={{fontSize: ".75rem"}} className="text-muted">Powered by <a href="https://www.pexels.com/">Pexels</a></p>
         </div>
       </div>
-    );
-  }
-}
+      <div style={{paddingTop: "95px"}} className="row m-0">
+        <div ref={gridParent} className="col-12 p-0 grid">
+          {photos.map(photo => (
+            <Photo
+              key={photo.id}
+              mediumImage={photo.src.medium}
+              largeImage={photo.src.large}
+              url={photo.url}
+              photographer={photo.photographer}
+            />
+          ))}
+        </div>
+      </div>
+    </div>
+  );
+};
 
 export default App;
